fix: added-tooltips
The Blockly-Block "Plot Diagram" used the tooltips of "Show Text/Number, which aren't helpful for users.

To make it more clear what exactly the different variables of the block define, I added tooltips in German and English.

Side note: I noticed and corrected a typo in 'display', which should make the title in the 'Plot Diagram' block now visible.